move the configs for the DataReader into rtkutils, too
